Import Mike Caulfield's Reader.htm Memex single-page-app.
"I want you to join us in our mission to create a federation that can function across servers like a net-enabled Memex. I want to see if you can come up with a better way to browse it, or to render it, or to whatever.I want you to join us in our mission to create a federation that can function across servers like a net-enabled Memex. I want to see if you can come up with a better way to browse it, or to render it, or to whatever."

And hopefully availability on the codewebs helps get a bump along that path.